Update time issues

+++ app/main/utils.py
@@ -103,49 +103,56 @@ def challenge_scheduler():
 	with app.app_context():
 
 		for user in db.session.query(User.uuid, User.timezoneoffset, User.email, User.username, User.receiveEmail,Tracker.last_activity).filter(User.uuid == Tracker.uuid ).all():
-			random_challenge = None
-			print(user)
-			currentuser = user.uuid
+			last_activity = user.last_activity
+			date_now = datetime.datetime.utcnow()
+			diff = last_activity - date_now
+			print('Diff: ' + str(diff))
+			if diff.minutes > 10:
+				random_challenge = None
+				print(user)
+				currentuser = user.uuid
 
-			# generates new random challenge
-			while True:
-				query = db.session.query(Challenge.cid)
-				challenges = query.all()
+				# generates new random challenge
+				while True:
+					query = db.session.query(Challenge.cid)
+					challenges = query.all()
 
-				# random challenge
-				random_challenge = random.choice(challenges)
+					# random challenge
+					random_challenge = random.choice(challenges)
 
-				activity = db.session.query(Activity.cid).filter(
-					Activity.user_id == currentuser, 
-					Activity.cid == random_challenge.cid).first()
-				
-				if activity is None:
-					break
+					activity = db.session.query(Activity.cid).filter(
+						Activity.user_id == currentuser, 
+						Activity.cid == random_challenge.cid).first()
+					
+					if activity is None:
+						break
 
-			# generates a unique id for activity.
-			activityid = (uuid.uuid4().int & (1<<29)-1)
+				# generates a unique id for activity.
+				activityid = (uuid.uuid4().int & (1<<29)-1)
 
-			#delete current activity
-			db.session.query(Activity).filter(
-				Activity.chal_status != 2, 
-				Activity.user_id == currentuser).delete()
-		
+				#delete current activity
+				db.session.query(Activity).filter(
+					Activity.chal_status != 2, 
+					Activity.user_id == currentuser).delete()
+			
 
-			# add a new activity for this user.
-			active = Activity(activity_id=activityid, 
-							  cid=random_challenge.cid,
-							  user_id=int(currentuser),
-							  current=True,
-							  chal_status=4)
+				# add a new activity for this user.
+				active = Activity(activity_id=activityid, 
+								  cid=random_challenge.cid,
+								  user_id=int(currentuser),
+								  current=True,
+								  chal_status=4)
 
-			db.session.add(active)
-			db.session.commit()
+				db.session.add(active)
+				db.session.commit()
 
-			# send new activity email
-			'''
-			if user and user.receiveEmail:
-				send_email(user)
-			'''
+				# send new activity email
+				'''
+				if user and user.receiveEmail:
+					send_email(user)
+				'''
+			else:
+				print('not yet')
 
 	print('Scheduler Executed')
 	sys.stdout.flush()
